// src/lib/api/generated/model/messageSummary.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Myelin Base RAG API
 * API documentation for Myelin Base RAG backend
 * OpenAPI spec version: 1.0.0
 */
import type { SourceDocument } from "./sourceDocument";
import type { FeedbackType } from "./feedbackType";

export interface MessageSummary {
  historyId: string;
  userQuery: string;
  aiResponse: string;
  sourceDocuments: SourceDocument[];
  feedback: FeedbackType;
  createdAt: string;
}